+++ src/imu/inertial_measurement_sensor.hpp
@@ -17,7 +17,7 @@ class InertialMeasurementSensor {
   virtual ~InertialMeasurementSensor() = default;
 
   explicit InertialMeasurementSensor(std::unique_ptr<i2c::I2C> i2c_handler) : i2c_handler_(std::move(i2c_handler)){};
-  auto Get(void) noexcept -> types::EuclideanVector<float>;
+  auto Get(void) noexcept -> types::EuclideanVector<int16_t>;
 
  protected:
   auto Mpu9255Detected(void) noexcept -> bool;
@@ -26,7 +26,7 @@ class InertialMeasurementSensor {
   auto ReadDataBytes(std::uint8_t read_from_register, std::uint16_t byte_size) noexcept -> std::pair<types::HalError, std::vector<std::uint8_t>>;
   auto ImuConnectionSuccessful(types::HalError imu_status) noexcept -> bool;
   /// Holds the local reference to euclidean sensor values
-  types::EuclideanVector<float> sensor_values_{-1, -1, -1};
+  types::EuclideanVector<int16_t> sensor_values_{-1, -1, -1};
   /// Holds the local reference to i2c handler
   std::unique_ptr<i2c::I2C> i2c_handler_;
   /// Holds value if Sensor was initialized correctly or not
